Rename as `escape-string`
