+++ rules/utils/escape-string.js
